+++ src/Frame.cpp
@@ -89,8 +89,8 @@ double Frame::getR() const {
     double sumY2 = 0;
     double sumXY = 0;
     for (uint32_t i = 0; i < _keyPoints.size(); ++i) {
-        int x = _keyPoints[i].pt.x;
-        int y = _keyPoints[i].pt.y;
+        int x = (int)_keyPoints[i].pt.x;
+        int y = (int)_keyPoints[i].pt.y;
         sumX += x;
         sumX2 += x * x;
         sumY += y;
